pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "labelforge"
version = "0.1.0"
description = "Deterministic, replayable, ablation-friendly multimodal labeling and synthetic-data pipeline"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "LabelForge Team" }
]
keywords = [
    "labeling",
    "multimodal",
    "vlm",
    "llm",
    "ray",
    "vllm",
    "synthetic-data",
    "machine-learning",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "ray[data]>=2.53.0",
    "vllm>=0.6.5",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "pyarrow>=15.0.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "orjson>=3.9.0",
    "rich>=13.0.0",
    "xxhash>=3.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.8.0",
    "ruff>=0.3.0",
    "pre-commit>=3.6.0",
]
s3 = [
    "boto3>=1.34.0",
]
gcs = [
    "google-cloud-storage>=2.14.0",
]

[project.scripts]
labelforge = "labelforge.cli:main"

[project.urls]
Homepage = "https://github.com/labelforge/labelforge"
Documentation = "https://github.com/labelforge/labelforge#readme"
Repository = "https://github.com/labelforge/labelforge"

[tool.hatch.build.targets.wheel]
packages = ["labelforge"]

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B017",   # pytest.raises without match
]

[tool.ruff.isort]
known-first-party = ["labelforge"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "ray.*",
    "vllm.*",
    "pyarrow.*",
    "xxhash.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "gpu: marks tests that require GPU",
]

[tool.coverage.run]
source = ["labelforge"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
